Optimized FileHelpers.deleteFileOrDirectory. Fixed a bug in RocksDBCache where close() was allowed to be executed more than once. If this second call would have happened after another instance with the same DB id was created, then this second call would wipe out that second instance's data.
Fixed a bug in StreamSegmentContainer.close() where two concurrent calls to this method could have adverse side effects (see RocksDBCache above).

Fixed some unit tests.

+++ common/src/main/java/com/emc/pravega/common/io/FileHelpers.java
@@ -1,7 +1,5 @@
 /**
- *
- *  Copyright (c) 2017 Dell Inc., or its subsidiaries.
- *
+ * Copyright (c) 2017 Dell Inc., or its subsidiaries.
  */
 package com.emc.pravega.common.io;
 
@@ -18,15 +16,11 @@ public class FileHelpers {
      * @return True if the target was deleted, false otherwise.
      */
     public static boolean deleteFileOrDirectory(File file) {
-        if (file.exists()) {
+        if (file.exists() && file.isDirectory()) {
             File[] files = file.listFiles();
             if (files != null) {
                 for (File f : files) {
-                    if (f.isDirectory()) {
-                        deleteFileOrDirectory(f);
-                    } else {
-                        f.delete();
-                    }
+                    deleteFileOrDirectory(f);
                 }
             }
         }
